requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.1
django-admin==2.0.2
django-excel-response2==3.0.5
django-six==1.0.5
excel-base==1.0.4
filelock==3.12.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
isoweek==1.3.3
jsonpickle==3.0.1
oauth2client==4.1.3
oauthlib==3.2.2
pipenv==2023.5.19
platformdirs==3.5.1
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.6.2
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
virtualenv-clone==0.5.7
xlwt==1.3.0
